+++ src/com/project/logic/Board.java
@@ -1,19 +1,211 @@
 package com.project.logic;
 
-public abstract class Board {
+public class Board {
 
-	public abstract void init();
-	
-	public abstract GridBoard copy();
-	
-	public abstract void print();
-	
-	public abstract void place(int stone, Point from, Point to);
-	
-	public abstract void push(Point from, Point to);
+//	if (yy > 0) 
+//		if (xx == 0) vertically down
+//		if (xx > 0) diagonal left top to right down
+//		if (xx < 0) diagonal right top to left down
+//		
+//	if (yy < 0) 
+//		if (xx == 0) vertically up
+//		if (xx > 0) diagonal left bottom to right top
+//		if (xx < 0) diagonal right bottom to left top
 
-	public abstract void move(Point from, Point to);
+	public static int GIPF_WHITE_VALUE = 11;
+	public static int GIPF_BLACK_VALUE = 22;
+	public static int WHITE_VALUE = 1;
+	public static int BLACK_VALUE = 2;
+	public static int BOARD_EDGE = -1;
+	public static int EMPTY_TILE = 0;
+	public static int VOID_TILE = -2;
+	
+	private int[][] grid;
 
-	public abstract boolean isEmpty(Point point);
+	public Board() {
+		this.init();
+	}
 
+	public Board(int[][] grid) {
+		this.grid = grid;
+	}
+
+	public void init() {
+		this.grid = new int[9][9];
+
+		for (int j = 8; j > 4; j--)
+			for (int i = 0; i < 4 - (8 - j); i++)
+				this.grid[i][j] = VOID_TILE;
+
+		for (int j = 0; j < 4; j++)
+			for (int i = 8; i > 4 + j; i--)
+				this.grid[i][j] = VOID_TILE;
+
+		for (int i = 4; i < 9; i++)
+			this.grid[i][8] = BOARD_EDGE;
+
+		for (int j = 8; j >= 4; j--)
+			this.grid[8][j] = BOARD_EDGE;
+
+		for (int i = 0; i < 5; i++)
+			this.grid[i][0] = BOARD_EDGE;
+
+		for (int j = 0; j < 5; j++)
+			this.grid[0][j] = BOARD_EDGE;
+
+		this.grid[5][1] = BOARD_EDGE;
+		this.grid[6][2] = BOARD_EDGE;
+		this.grid[7][3] = BOARD_EDGE;
+
+		this.grid[1][5] = BOARD_EDGE;
+		this.grid[2][6] = BOARD_EDGE;
+		this.grid[3][7] = BOARD_EDGE;
+
+		// placing the stones
+		this.grid[1][1] = WHITE_VALUE;
+		this.grid[4][1] = BLACK_VALUE;
+		this.grid[7][4] = WHITE_VALUE;
+		this.grid[7][7] = BLACK_VALUE;
+		this.grid[4][7] = WHITE_VALUE;
+		this.grid[1][4] = BLACK_VALUE;
+	}
+
+	public Board copy() {
+		int[][] gridCopy = new int[9][9];
+
+		for (int i = 0; i < gridCopy.length; i++)
+			for (int j = 0; j < gridCopy[0].length; j++)
+				gridCopy[i][j] = this.grid[i][j];
+
+		return new Board(gridCopy);
+	}
+
+	public void print() {
+		for (int j = 0; j < this.grid[0].length; j++) {
+			for (int i = 0; i < this.grid.length; i++) {
+				if (this.grid[i][j] > -1) System.out.print(" ");
+				System.out.print(this.grid[i][j] + " ");
+			}
+			System.out.println();
+		}
+	}
+
+	public void place(int stone, Point from, Point to) {
+		this.push(from, to);
+		this.grid[to.getX()][to.getY()] = stone;
+	}
+
+	public void push(Point from, Point to) {
+		int x1 = from.getX();
+		int x2 = to.getX();
+		int y1 = from.getY();
+		int y2 = to.getY();
+		int xx = (int) (x2 - x1);
+		int yy = (int) (y2 - y1);
+
+		if (!isEmpty(new Point(x2, y2))) push(to, new Point(x2 + xx, y2 + yy));
+		if (grid[x1][y1] > 0) move(from, to);
+	}
+
+	public void move(Point from, Point to) {
+		int x1 = from.getX();
+		int x2 = to.getX();
+		int y1 = from.getY();
+		int y2 = to.getY();
+		this.grid[x2][y2] = this.grid[x1][y1];
+		this.grid[x1][y1] = 0;
+	}
+
+	public boolean isEmpty(Point point) {
+		return this.grid[(int) point.getX()][(int) point.getY()] <= 0;
+	}
+	
+	public void checkForLines() {
+		int lineStartX = -1, lineStartY = -1, lineEndX = -1, lineEndY = -1;
+
+		int counter = -1;
+		int prevValue = -1;
+
+		// checks for vertical lines
+		for (int i = 1; i < grid.length - 1; i++) {
+			for (int j = 1; j < grid[i].length - 1; j++) {
+				if (prevValue == grid[i][j] && prevValue > 0) counter++;
+				else {
+					counter = 1;
+					prevValue = grid[i][j];
+				}
+				if (counter == 4) {
+					lineEndX = i;
+					lineEndY = j;
+					lineStartX = i;
+					lineStartY = j - 3;
+				}
+			}
+			counter = 0;
+			prevValue = -1;
+		}
+
+		// checks for left down right up lines
+		for (int j = 0; j < grid[0].length; j++) {
+			for (int i = 0; i < grid.length; i++) {
+				if (prevValue == grid[i][j] && prevValue > 0) counter++;
+				else {
+					counter = 1;
+					prevValue = grid[i][j];
+				}
+				if (counter == 4) {
+					lineEndX = i;
+					lineEndY = j;
+					lineStartX = i - 3;
+					lineStartY = j;
+				}
+			}
+			counter = 1;
+			prevValue = -1;
+		}
+
+		// checks for lines left top right bottom
+		for (int j = 3; j >= 0; j--) {
+			for (int i = 1; i < (9 - j) - 1; i++) {
+				if (prevValue == grid[i][j + i] && prevValue > 0) counter++;
+				else {
+					counter = 1;
+					prevValue = grid[i][j + i];
+				}
+				if (counter == 4) {
+					lineEndX = i;
+					lineEndY = j + i;
+					lineStartX = i - 3;
+					lineStartY = j - 3 + i;
+				}
+			}
+			counter = 0;
+			prevValue = -1;
+		}
+
+		for (int i = 1; i < 4; i++) {
+			for (int j = 1; j < (9 - i) - 1; j++) {
+				if (prevValue == grid[i + j][j] && prevValue > 0) counter++;
+				else {
+					counter = 1;
+					prevValue = grid[i + j][j];
+				}
+				if (counter == 4) {
+					lineEndX = i + j;
+					lineEndY = j;
+					lineStartX = i - 3 + j;
+					lineStartY = j - 3;
+				}
+			}
+			counter = 0;
+			prevValue = -1;
+		}
+
+		System.out.println(new Point(lineStartX, lineStartY) + " " + new Point(lineEndX, lineEndY));
+	}
+
+
+//	public int[][] getGrid() {
+//		return this.grid;
+//	}
 }
