+++ src/com/project/visuals/ResourceLoader.java
@@ -10,6 +10,7 @@ public class ResourceLoader {
 
 	public static Image WHITE_STONE, BLACK_STONE, OUTER_DOT;
 	public static Image MENU_ICON, SOUND_ICON, SOUND_MUTED_ICON, SETTINGS_ICON, HELP_ICON;
+	public static Image ICON;
 	
 	public void init() throws IOException {
 		WHITE_STONE = ImageIO.read(this.getResource("res/icons/WhiteStone.png"));
@@ -21,6 +22,8 @@ public class ResourceLoader {
 		SOUND_MUTED_ICON = ImageIO.read(this.getResource("res/icons/MutedVolume48.png"));
 		SETTINGS_ICON = ImageIO.read(this.getResource("res/icons/Settings48.png"));
 		HELP_ICON = ImageIO.read(this.getResource("res/icons/QuestionMark48.png"));
+		
+		ICON = ImageIO.read(this.getResource("res/icons/Icon.png"));
 	}
 	
 //	public InputStream getResource(String path) {
